put all in class
